fix: change invocation of mysqli::real_connect, ssl_set based on flags
Flagged by PHP 8.

For `real_connect()`, flags MUST be an integer, and cannot be null, once we hit PHP 8.
As such, this patch modifies the mysqli connection class to vary how it calls the method based on whether or not `$flags` is null.

For `ssl_set()`, all arguments must be strings, not null, once we hit PHP 8.
As such, this patch modifies the arguments to ssl_set to default to empty strings instead of null values.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>